backend/requirements.txt
requests
boto3
Flask
Flask-RESTful
dropbox
google-api-core
gcloud
azure-storage-blob